2009-08-04  Jeffrey Stedfast  <fejj@novell.com>
	* utils.cpp (write_all): Changed to just return int (-1 for fail,
	0 for success).
	(ExtractFile): Don't fsync(). Saves us a lot of time.


svn path=/trunk/moon/; revision=139355

+++ src/utils.h
@@ -77,7 +77,7 @@ int RemoveDir (const char *dir);
 
 int CopyFileTo (const char *filename, int fd);
 
-ssize_t write_all (int fd, char *buf, size_t len);
+int write_all (int fd, char *buf, size_t len);
 
 cairo_t *measuring_context_create (void);
 void     measuring_context_destroy (cairo_t *cr);
